splitting out stale handler so its easy to remove later
