+++ example/modules/Optimized.js
@@ -2,8 +2,6 @@ import React, {Component, PropTypes} from 'react';
 import TypeSize from './TypeSize';
 import {Row, Column} from '../../src';
 
-let i = 1;
-
 /*
  * Example of what happens if someone uses shouldComponentUpdate
  */
@@ -13,7 +11,6 @@ export default class Optimized extends Component {
   }
 
   render() {
-    console.log('render', i++);
     return (
       <Row>
         <Column width="1/2">
